+++ com_opensim/admin/views/misc/view.html.php
@@ -1,7 +1,7 @@
 <?php
 /*
  * @component jOpenSim
- * @copyright Copyright (C) 2017 FoTo50 http://www.jopensim.com/
+ * @copyright Copyright (C) 2018 FoTo50 http://www.jopensim.com/
  * @license GNU/GPL v2 http://www.gnu.org/licenses/gpl-2.0.html
  */
 // no direct access
@@ -9,16 +9,25 @@
 defined( '_JEXEC' ) or die( 'Restricted access' );
  
 jimport( 'joomla.application.component.view');
+JLoader::register('jOpenSimHelper', JPATH_COMPONENT.'/helpers/jopensimhelper.php');
+JTable::addIncludePath(JPATH_COMPONENT.DIRECTORY_SEPARATOR.'tables');
  
 class opensimViewmisc extends JViewLegacy {
 	public function display($tpl = null) {
 		JHTML::_('behavior.modal');
+//		$this->state		= $this->get('State');
 		$document			= JFactory::getDocument();
 		$document->addStyleSheet(JURI::base(true).'/components/com_opensim/assets/opensim.css');
 		$this->model		= $this->getModel('misc');
-		$settings			= $this->model->getSettingsData();
+		$this->settings			= $this->model->getSettingsData();
+		if($this->settings['remoteadminsystem'] == "multiple") {
+			$this->connectedsimulators = $this->model->getconnectedsimulators();
+		} else {
+			$this->connectedsimulators = null;
+		}
 		$task				= JFactory::getApplication()->input->get( 'task', '', 'method', 'string');
 		$this->sidebar		= JHtmlSidebar::render();
+		$this->canDo		= jOpenSimHelper::getActions();
 
 		$assetinfo			= pathinfo(JPATH_COMPONENT_ADMINISTRATOR);
 		$assetpath			= "components".DIRECTORY_SEPARATOR.$assetinfo['basename'].DIRECTORY_SEPARATOR."assets".DS;
@@ -27,21 +36,31 @@ class opensimViewmisc extends JViewLegacy {
 
 		switch($task) {
 			case "getopensimversion":
-				// $remotehost			= $settings['remotehost'];
-				// $this->remotehost	= $remotehost;
+				$this->simulators	= $this->model->getSimulators();
+				foreach($this->simulators AS $key => $simulator) {
+					if($simulator['connected'] === FALSE) unset($this->simulators[$key]);
+				}
 				$tpl				= "getopensimversion";
 			break;
 			case "addregion":
-				$remotehost			= $settings['remotehost'];
+				$remotehost			= $this->settings['remotehost'];
 				$this->remotehost	= $remotehost;
+				$this->simulators	= $this->model->getSimulators();
+				foreach($this->simulators AS $key => $simulator) {
+					if($simulator['connected'] === FALSE) unset($this->simulators[$key]);
+				}
 				$tpl				= "addregion";
 			break;
 			case "sendmessage":
+				$this->simulators	= $this->model->getSimulators();
+				foreach($this->simulators AS $key => $simulator) {
+					if($simulator['connected'] === FALSE) unset($this->simulators[$key]);
+				}
 				$tpl				= "sendmessage";
 			break;
 			case "terminals":
 				$terminalList		= $this->model->getTerminalList(1);
-		 		$pagination			=& $this->get('Pagination');
+		 		$pagination			= $this->get('Pagination');
 				$this->pagination	= $pagination;
 				$this->terminalList	= $terminalList;
 				$tpl				= "terminals";
@@ -65,18 +84,29 @@ class opensimViewmisc extends JViewLegacy {
 				$this->pingAnswer	= $pingAnswer;
 				$tpl				= "terminalping";
 			break;
+			case"managesimulators":
+				$this->simulators	= $this->model->getSimulators();
+				$tpl				= "managesimulators";
+			break;
 			default:
-				if ($settings['enableremoteadmin'] == "1") { 
-                    $misclinks['addregion']         = "<a href='index.php?option=com_opensim&view=misc&task=addregion'>".JText::_('JOPENSIM_ADDREGION')."</a>";
-					$misclinks['sendmessage']       = "<a href='index.php?option=com_opensim&view=misc&task=sendmessage'>".JText::_('SENDGLOBALMESSAGE')."</a>";
-                    $misclinks['getopensimversion'] = "<a href='index.php?option=com_opensim&view=misc&task=getopensimversion'>".JText::_('GETOPENSIMVERSION')."</a>";
-                } else {
-                    $misclinks['addregion']         = JText::_('JOPENSIM_ADDREGION')." (".JText::_('DISABLED_NOREMOTEADMIN').")";
-					$misclinks['sendmessage']       = JText::_('SENDGLOBALMESSAGE')." (".JText::_('DISABLED_NOREMOTEADMIN').")";
-                    $misclinks['getopensimversion'] = JText::_('GETOPENSIMVERSION')." (".JText::_('DISABLED_NOREMOTEADMIN').")";
+				$document->addStyleSheet(JURI::base(true).'/components/com_opensim/assets/quickiconstyle.css?v=2.6.8');
+				if($this->canDo->get('core.remoteadmin')) {
+					if ($this->settings['enableremoteadmin'] == "1") {
+	                    $misclinks['addregion']       = $this->renderButton('index.php?option=com_opensim&view=misc&task=addregion','icon-48-region.png',JText::_('JOPENSIM_ADDREGION'));
+//	                    $misclinks['addregion']         = "<a href='index.php?option=com_opensim&view=misc&task=addregion'>".JText::_('JOPENSIM_ADDREGION')."</a>";
+	                    $misclinks['sendmessage']       = $this->renderButton('index.php?option=com_opensim&view=misc&task=sendmessage','icon-48-sendmessage.png',JText::_('JOPENSIM_SENDMESSAGE'),JText::_('JOPENSIM_SENDMESSAGE_DESC'));
+//						$misclinks['sendmessage']       = "<a href='index.php?option=com_opensim&view=misc&task=sendmessage'>".JText::_('SENDGLOBALMESSAGE')."</a>";
+	                    $misclinks['getopensimversion'] = $this->renderButton('index.php?option=com_opensim&view=misc&task=getopensimversion','icon-48-os-osversion.png',JText::_('GETOPENSIMVERSION'),JText::_('GETOPENSIMVERSION_DESC'));
+//	                    $misclinks['getopensimversion'] = "<a href='index.php?option=com_opensim&view=misc&task=getopensimversion'>".JText::_('GETOPENSIMVERSION')."</a>";
+	                } else {
+	                    $misclinks['addregion']         = JText::_('JOPENSIM_ADDREGION')." (".JText::_('DISABLED_NOREMOTEADMIN').")";
+						$misclinks['sendmessage']       = JText::_('SENDGLOBALMESSAGE')." (".JText::_('DISABLED_NOREMOTEADMIN').")";
+	                    $misclinks['getopensimversion'] = JText::_('GETOPENSIMVERSION')." (".JText::_('DISABLED_NOREMOTEADMIN').")";
+					}
 				}
-				if($settings['addons'] & 8) {
-					$misclinks['terminals']		= "<a href='index.php?option=com_opensim&view=misc&task=terminals'>".JText::_('MANAGETERMINALS')."</a>";
+				if($this->settings['addons'] & 8) {
+                    $misclinks['terminals']       = $this->renderButton('index.php?option=com_opensim&view=misc&task=terminals','icon-48-os-terminal.png',JText::_('JOPENSIM_TERMINALS'),JText::_('JOPENSIM_TERMINALS_DESC'));
+//					$misclinks['terminals']		= "<a href='index.php?option=com_opensim&view=misc&task=terminals'>".JText::_('MANAGETERMINALS')."</a>";
 				}
 			break;
 		}
@@ -94,28 +124,55 @@ class opensimViewmisc extends JViewLegacy {
 			case "getopensimversion":
                 // JToolBarHelper::publish('getopensimulatorversion');
 				JToolBarHelper::cancel('canceladdregion','JCANCEL');
+				if($this->canDo->get('core.simulators')) {
+					JToolbarHelper::apply('managesimulators','JOPENSIM_SIMULATORS');
+				}
 				JToolBarHelper::help("", false, JText::_('JOPENSIM_HELP_MISC_ADDREGION'));
 			break;
 			case "addregion":
-				JToolBarHelper::save('createregionsend');
+				JToolBarHelper::addnew('createregionsend','JOPENSIM_ADDREGION');
 				JToolBarHelper::cancel('canceladdregion','JCANCEL');
+				if($this->canDo->get('core.simulators')) {
+					JToolbarHelper::apply('managesimulators','JOPENSIM_SIMULATORS');
+				}
 				JToolBarHelper::help("", false, JText::_('JOPENSIM_HELP_MISC_ADDREGION'));
 			break;
 			case "sendmessage":
 				JToolBarHelper::publish('sendoutmessage');
 				JToolBarHelper::cancel('cancelmessage','JCANCEL');
+				if($this->canDo->get('core.simulators')) {
+					JToolbarHelper::apply('managesimulators','JOPENSIM_SIMULATORS');
+				}
 				JToolBarHelper::help("", false, JText::_('JOPENSIM_HELP_MISC_SENDMESSAGE'));
 			break;
 			case "terminals":
-				JToolBarHelper::deleteList(JText::_('DELETETERMINALSURE'),"deleteTerminal",JText::_('DELETETERMINAL'),true,false);
-				JToolBarHelper::editList('terminaledit');
+				if($this->canDo->get('core.delete')) {
+					JToolBarHelper::deleteList(JText::_('DELETETERMINALSURE'),"deleteTerminal",JText::_('DELETETERMINAL'),true,false);
+				}
+				if($this->canDo->get('core.edit')) {
+					JToolBarHelper::editList('terminaledit');
+				}
 				JToolBarHelper::cancel('cancelTerminal','JCANCEL');
 			break;
 			case "terminaledit":
-				JToolBarHelper::save('saveTerminal');
+				if($this->canDo->get('core.edit')) {
+					JToolBarHelper::save('saveTerminal');
+				}
 				JToolBarHelper::cancel('terminals');
 			break;
+			case "managesimulators":
+				if($this->canDo->get('core.edit')) {
+					JToolBarHelper::save('saveSimulators');
+				}
+				JToolBarHelper::cancel('cancelSimulators','JCANCEL');
+				if($this->canDo->get('core.delete')) {
+					JToolBarHelper::deleteList(JText::_('JOPENSIM_SIMULATORS_DELETESURE'),"deleteSimulator",JText::_('JOPENSIM_SIMULATORS_DELETE'),true,true);
+				}
+			break;
 			default:
+				if($this->canDo->get('core.simulators')) {
+					JToolbarHelper::apply('managesimulators','JOPENSIM_SIMULATORS');
+				}
 				$os_settings = $this->model->getSettingsData();
 				if(isset($os_settings['remoteadmin_enabled']) && $os_settings['remoteadmin_enabled'] == 1) {
 					JToolBarHelper::custom("sendmessage","osmisc","opensim",JText::_('SENDMESSAGE2USER'),false,false);
@@ -126,6 +183,26 @@ class opensimViewmisc extends JViewLegacy {
 				JToolBarHelper::help("", false, JText::_('JOPENSIM_HELP_MISC'));
 			break;
 		}
+	}
+
+	public function getHostFrom($url) {
+		return parse_url($url, PHP_URL_HOST);
+	}
+
+	public function getPortFrom($url) {
+		return parse_url($url, PHP_URL_PORT);
+	}
+
+	public function renderButton($link,$image,$text,$title = "") {
+		if(!$title) $title = $text;
+		$params = array('title'=>$title, 'border'=>'0');
+		$button  = "<div class='icon-wrapper'>";
+		$button .= "<div class='icon'>";
+		$button .= sprintf("<a href='%s' class='os_mainscreen'>",$link);
+		$button .= JHTML::_('image', 'administrator/components/com_opensim/assets/images/'.$image,$title,$params);
+		$button .= sprintf("<span>%s</span></a>",$text);
+		$button .= "</div></div>\n";
+		return $button;
 	}
 }
 ?>
